Small fixes for Monsters - no release!
Fixed Light armoured ogre not showing up as an option when having 3 Ogres in the unit.

This is pending further changes before I can release this as the Allied/Bound/Wild rules from previous versions are incorrect. The allied/bound/wild cost is a one-off thing and does not affect the actual unit cost.

Removed the word monster from the allied/bound/wild rule name as that has intrinsic meaning in the rule set.